Refactor home contaner component
